pro-tax-ai-chatbot/requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
numpy
streamlit
groq